fix: Fix performance test imports and add 18/32 controller cases
- Move shared utilities from conftest.py to helpers.py to fix
  ModuleNotFoundError in CI (conftest is auto-loaded by pytest,
  not importable as a module)
- Add 18 and 32 controller test cases (common real-world counts)
- Bump memory stress test from 20 to 32 controllers
